+++ src/UserInterface.cpp
@@ -60,8 +60,8 @@ void UI::Buttons_Init() {
 
 void UI::Buttons_Unload()
 {
-	int sz = buttonTest.size();
-	for (int i = 0; i < sz; ++i) {
+	size_t sz = buttonTest.size();
+	for (size_t i = 0; i < sz; ++i) {
 		buttonTest.pop_back();
 	}
 }
